Move ResourceType from resource to pages app and rename to RelatedLinkType
