+++ cdhweb/pages/models.py
@@ -250,3 +250,17 @@ class PageIntro(models.Model):
 
     def __str__(self):
         return self.page.title
+
+
+class RelatedLinkType(models.Model):
+    '''Resource type for associating particular kinds of URLs
+    with people and projects (e.g., project url, GitHub, Twitter, etc)'''
+    name = models.CharField(max_length=255)
+    sort_order = models.PositiveIntegerField(default=0, blank=False,
+                                             null=False)
+
+    class Meta:
+        ordering = ['sort_order']
+
+    def __str__(self):
+        return self.name
